requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.51.2
dvc-data==3.15.2
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.112.0
filelock==3.19.1
Flask==3.1.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gto==1.8.0
gunicorn==22.0.0
h11==0.16.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==7.2.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
kombu==5.5.4
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mlflow==2.14.1
multidict==6.6.4
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==3.11.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
py4j==0.10.9.7
pyarrow==15.0.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
pyspark==3.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
querystring-parser==1.2.4
requests==2.32.3
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.1
scipy==1.16.1
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.37.2
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.5
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
